requirements.txt
aiogram==3.3.0
gspread==6.0.0
google-api-python-client==2.126.0
google-auth==2.29.0
openpyxl==3.1.2
requests==2.32.3
